fix role and field validation in items routes

+++ api/controllers/items/deleteItem.js
@@ -8,8 +8,10 @@ const Item = require('../../models/Item');
 const deleteItem = (req, res) => {
   const { id } = req.params;
 
-  // Verify Roles
-  verifyRole(req.user, res);
+  // Verify roles
+  if (!verifyRole(req.user)) {
+    return res.status(401).json({ msg: 'You are not authorized to do this.' });
+  }
 
   Item.findOneAndRemove({ _id: id })
     .then((removedItem) => {
